Create commands so we can do multiple things in our single binary, right now we only serve
